Seperating Ingest and Connector logic
